Add CatTag in HomeScreen, set scroll to 0 when changing category

+++ app/src/main/java/pixel/breakfast/data/dto/MenuItemResp.kt
@@ -8,6 +8,6 @@ import kotlinx.serialization.Serializable
 data class MenuItemResp(
     @SerialName("id") val id: String,
     @SerialName("category") val category: String,
-    @SerialName("price") val price: Int,
+    @SerialName("price") val price: Double,
     @SerialName("img_res") val imgRes: String,
 )
